Introduces Transport to move data between buffers.
BufferBuilder now allows constructing various types of buffers by
letting a user specify the element type and some initial data.

It uses Transport which abstracts over the various primitive array
types and List for reference types.

Doing bulk reads and writes required a concept of a slice of data
so reparameterized buffers and slices to have an element type and
a slice type.

Sketched out a channel which combines a write-only sink and a
read-only source.